+++ surfalize/file/common.py
@@ -47,13 +47,43 @@ def get_unit_conversion(from_unit, to_unit):
     factor: float
         Factor by which to multiply the original values.
     """
+    from_unit = _sanitize_mu(from_unit)
     if from_unit not in UNIT_EXPONENT or to_unit not in UNIT_EXPONENT:
         raise ValueError('Unit does not exist.')
-    from_unit = _sanitize_mu(from_unit)
     exponent = UNIT_EXPONENT[from_unit] - UNIT_EXPONENT[to_unit]
     return 10**exponent
 
-def read_binary_layout(filehandle, layout, fast=True):
+def write_binary_layout(filehandle, layout, data, encoding='utf-8'):
+    """
+    Writes a binary layout to a file.
+
+    Parameters
+    ----------
+    filehandle: file object
+        File-like object to read the data from.
+    layout: tuple[tuple[str, str, bool] | tuple[None, int, None]]
+        Layout of the bytes to read as a tuple of tuples in the form (<name>, <format>, <skip_fast>) or
+        (None, <n_bytes>, None) for reserved bytes.
+    data: dict[str: any]
+        Dictionary containing keys that correspond to the name value in the layout tuple and the values to write
+        to the file as keys.
+    encoding: str, Default utf-8
+            Encoding of characters in the file. Defaults to utf-8.
+
+    Returns
+    -------
+    None
+    """
+    for name, format_, _ in layout:
+        if name is None:
+            filehandle.write(b'\x00' * format_)
+            continue
+        value = data[name]
+        if isinstance(value, str):
+            value = value.encode(encoding)
+        filehandle.write(struct.pack(format_, value))
+
+def read_binary_layout(filehandle, layout, fast=True, encoding='utf-8'):
     """
     Reads a binary layout specified by a tuple of tuples from an opened file and returns a dict with the read values.
     The layout must be provided in the form:
@@ -80,6 +110,8 @@ def read_binary_layout(filehandle, layout, fast=True):
         (None, <n_bytes>, None) for reserved bytes.
     fast: bool, default True
         If true, skips all values that are marked for skipping in fast mode.
+    encoding: str, Default utf-8
+            Encoding of characters in the file. Defaults to utf-8.
 
     Returns
     -------
@@ -96,7 +128,7 @@ def read_binary_layout(filehandle, layout, fast=True):
             continue
         unpacked_data = struct.unpack(f'{format}', filehandle.read(size))[0]
         if isinstance(unpacked_data, bytes):
-            unpacked_data = unpacked_data.decode().strip()
+            unpacked_data = unpacked_data.decode(encoding).rstrip(' \x00')
         result[name] = unpacked_data
     return result