+++ surfalize/roughness/hybrid.pyx
@@ -6,7 +6,7 @@ ctypedef fused sfloat:
     float
     double
 
-cdef sfloat triangle_area(sfloat x0, sfloat y0, sfloat z0, sfloat x1, sfloat y1, sfloat z1):
+cdef sfloat triangle_area(sfloat x0, sfloat y0, sfloat z0, sfloat x1, sfloat y1, sfloat z1) nogil:
     """
     Calculates the area of a 3d triangle spanned by the points A, B, C based on the two spanning vectors
     AB and AC, where AB = (x0, y0, z0) and AC = (x1, y1, z1). The area is calculated by computing the
@@ -29,9 +29,10 @@ def surface_area(sfloat[:,:] p, sfloat dx, sfloat dy):
     cdef Py_ssize_t imax = p.shape[0] - 1
     cdef Py_ssize_t jmax = p.shape[1] - 1
     cdef double a1, a2
-    for i in range(imax):
-        for j in range(jmax):
-            a1 = triangle_area(0, dy, p[i+1,j] - p[i,j], dx, 0, p[i,j+1] - p[i,j])
-            a2 = triangle_area(dx, 0, p[i+1,j] - p[i+1,j+1], 0, dy, p[i,j+1] - p[i+1,j+1])
-            total_area += a1 + a2
+    with nogil:
+        for i in range(imax):
+            for j in range(jmax):
+                a1 = triangle_area(0, dy, p[i+1,j] - p[i,j], dx, 0, p[i,j+1] - p[i,j])
+                a2 = triangle_area(dx, 0, p[i+1,j] - p[i+1,j+1], 0, dy, p[i,j+1] - p[i+1,j+1])
+                total_area += a1 + a2
     return total_area